Switch to Dub and DMD
Switched to the Dub build tool and the DMD compiler, since that seems to
work better and produce a much smaller statically linked binary.

Switch to Dub with LDC in GH Actions

GDC doesn't seem to link the libraries correctly, so I'm trying Dub with LDC instead to see if that produces a static binary.

Update d.yml

Switch to building with Dub and DMD

Switched to more standard tools for building and working with D code.
Also DMD seems to produce a MUCH smaller binary which is very nice, but
I have to make sure it's statically linked.