Flag precedence, Improve dirsizes, fix --rm -r
Re-ordered the way that flags and options are handled so that they
follow the correct precedence ordering.

Improved the directorysizes file handling. Now deleted/restored files
are properly removed and the folder size is actually calculated
correctly. Along with some formatting fixes

Fixed an issue with --rm where it couldn't delete directories even with -r.