Refactor into Multiple files
I resisted doing this for a while since I wanted to keep the project
small and ideally keep everything in a single file. But the previous
addition of automated tests meant things were broken up anyway, so I
decided to embrace it and spread things out in the hope that it is now a
bit more readable.

I also discovered that dub has a lint command, so I ran that and fixed
the issues as well as enabled it in GH actions.