--orphans command for finding lost files
Files in the trash that do not have a matching .trashinfo will not show
up in --list. This is per-spec, and is roughly how other implementations
work. But that means files can be lurking in the trash directory and
taking up space.

The new --orphans flag displays these so that the user knows they are
there and can deal with them.

Add basic test for orphans

Going to come back to this and the test for --list and do
something more robust