v20 Maintenance release
This is a minor maintenance release focused on
improving the tooling and build process.

I apologize to any package maintainers that are
affected, and hope you agree with some of the
changes!

Tooling changes:
- Copyright info was updated
- The Rakefile has been replaced with a Justfile
- Ronn-ng has been replaced with scdoc

With these the build dependency on Ruby is now
gone!
The package.rb file remains to build DEBs
and RPMs with FPM. However I will no longer be
providing these myself. I don't think users
should be relying on these.

Code changes:
- Functions were all annotated with @safe or
  @trusted as applicable
- Version name information is again in dub.json

My hope is that by relying on D's safety features
we can make stronger guarantees about the program,
which helps long-term maintenance.

The version is now 20 and the version name is now
"Annette Birkin" from Resident Evil 2.